+++ src/pred_bench/pred_benchmark_job.h
@@ -29,7 +29,7 @@
 
 typedef struct {
     int call_index;
-    long msize;
+    size_t msize;
     long n_rep;
 } job_t;
